Splitting behaviour on tags (#2)
Two changes are introduced to the vue-input-tag:

Rather than listening for the keycode pressed, we listen for the key pressed.
This was due to two issues:
the fact that keycodes weren't alike across platforms. For instance the space ' ' does not resolve to the same keycode on MacOS (188) and Linux (222) (Windows not tested).
the fact that the danish character 'æ' does resolve to the same keycode as ';' (semicolon, keycode 186) does; since we were listening for keycode 186 as a means to separate tags it has the unfortunate effect that danish names such as 'Lærke' and 'Æske' could not be input.
The ability to provide a custom (regex) pattern used when splitting copy-pasted text. The need for having a custom pattern arose for when entering emails; since a) emails cannot contain spaces and b) more importantly, we wish to allow the user to split input using a space when entering emails.